src/hydra_core/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via django
astroid==2.8.4
    # via pylint
attrs==21.2.0
    # via pytest
black==21.10b0
    # via -r requirements.in
click==8.0.3
    # via black
django==3.2.9
    # via
    #   -r requirements.in
    #   djangorestframework
django-environ==0.8.1
    # via -r requirements.in
django-rest-framework==0.1.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via django-rest-framework
factory-boy==3.2.1
    # via -r requirements.in
faker==9.8.0
    # via factory-boy
flake8==4.0.1
    # via -r requirements.in
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via
    #   -r requirements.in
    #   pylint
lazy-object-proxy==1.6.0
    # via astroid
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
mypy==0.910
    # via -r requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.2
    # via pytest
paste==3.5.0
    # via -r requirements.in
pathspec==0.9.0
    # via black
platformdirs==2.4.0
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pylint==2.11.1
    # via
    #   -r requirements.in
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.4.4
    # via -r requirements.in
pylint-plugin-utils==0.6
    # via pylint-django
pyparsing==2.4.7
    # via packaging
pytest==6.2.5
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.4.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
pytz==2021.3
    # via django
regex==2021.11.2
    # via black
six==1.16.0
    # via
    #   paste
    #   python-dateutil
sqlparse==0.4.2
    # via django
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   mypy
    #   pylint
    #   pytest
tomli==1.2.2
    # via black
types-waitress==2.0.1
    # via -r requirements.in
typing-extensions==3.10.0.2
    # via
    #   astroid
    #   black
    #   mypy
    #   pylint
waitress==2.0.0
    # via -r requirements.in
whitenoise==5.3.0
    # via -r requirements.in
wrapt==1.13.3
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
